requirements.txt
amqp==1.4.9
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.7.1
billiard==3.3.0.23
bs4==0.0.1
cachetools==3.0.0
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.3.0
Django==2.1.5
django-celery==3.2.2
django-celery-beat==1.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-redis==4.10.0
django-timezone-field==3.0
djangorestframework==3.9.1
ez-setup==0.9
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gspread==3.1.0
html5==0.0.9
httplib2==0.12.0
idna==2.8
ipinfo-django==0.1.4
itsdangerous==1.1.0
Jinja2==2.10
kombu==3.0.37
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
MechanicalSoup==0.11.0
more-itertools==5.0.0
oauth2client==4.1.3
passlib==1.7.1
Pillow==5.4.1
pluggy==0.8.1
prettytable==0.7.2
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pytest==4.1.1
python-crontab==2.3.6
python-dateutil==2.7.5
pytz==2018.9
redis==3.0.1
requests==2.21.0
rsa==4.0
selenium==3.141.0
six==1.12.0
soupsieve==1.7.1
SQLAlchemy==1.2.16
stripe==2.18.1
urllib3==1.24.1
validators==0.12.4
vine==1.2.0
virtualenv==16.2.0
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
